Week6/Langfuse-Guardrails-Task/requirements.txt
# Core LangChain
langchain>=0.3.0
langchain-aws>=0.2.0
langchain-core>=0.3.0

# AWS Bedrock
boto3>=1.34.0

# LangFuse (tracing & monitoring)
langfuse>=2.0.0

# Guardrails AI
guardrails-ai>=0.5.0

# RAG System
chromadb==0.4.22
sentence-transformers>=2.7.0
huggingface-hub>=0.20.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0

# Web Search
ddgs>=6.0.0

# Document Processing
PyPDF2==3.0.0

# Utilities
python-dotenv==1.0.0
pandas>=2.0.0
